Moved NotificationProvider interface into its own file
