+++ rover/build.gradle
@@ -13,7 +13,7 @@ ext {
     siteUrl = 'https://github.com/RoverPlatform/rover-android'
     gitUrl = 'https://github.com/RoverPlatform/rover-android.git'
 
-    libraryVersion = '0.1.0'
+    libraryVersion = '0.2.1'
 
     developerId = 'ata-n'
     developerName = 'Ata N'
